Clean up influxdb.
