+++ stats_test.go
@@ -2,20 +2,20 @@ package influxdb_test
 
 import (
 	"testing"
-	"time"
 
 	"github.com/influxdb/influxdb"
 )
 
-func TestStats_SetAndGet(t *testing.T) {
+// Ensure the stats can set and retrieve a value for a key.
+func TestStats_Set(t *testing.T) {
 	s := influxdb.NewStats("foo")
-
 	s.Set("a", 100)
 	if s.Get("a") != 100 {
 		t.Fatalf("stats set failed, expected 100, got %d", s.Get("a"))
 	}
 }
 
+// Ensure that stats can add a value to a key.
 func TestStats_Add(t *testing.T) {
 	s := influxdb.NewStats("foo")
 
@@ -25,6 +25,17 @@ func TestStats_Add(t *testing.T) {
 	}
 }
 
+// Ensure that stats can subtract by adding a negative number.
+func TestStats_Add_Negative(t *testing.T) {
+	s := influxdb.NewStats("foo")
+
+	s.Add("a", -200)
+	if s.Get("a") != -200 {
+		t.Fatalf("stats set failed, expected -200, got %d", s.Get("a"))
+	}
+}
+
+// Ensure that stats can increment a value by 1.
 func TestStats_Inc(t *testing.T) {
 	s := influxdb.NewStats("foo")
 
@@ -37,15 +48,6 @@ func TestStats_Inc(t *testing.T) {
 	s.Inc("b")
 	if s.Get("b") != 1 {
 		t.Fatalf("stats Inc failed, expected 1, got %d", s.Get("b"))
-	}
-}
-
-func TestStats_AddNegative(t *testing.T) {
-	s := influxdb.NewStats("foo")
-
-	s.Add("a", -200)
-	if s.Get("a") != -200 {
-		t.Fatalf("stats set failed, expected -200, got %d", s.Get("a"))
 	}
 }
 
@@ -78,55 +80,24 @@ func TestStats_Diff(t *testing.T) {
 }
 
 func TestStats_Snapshot(t *testing.T) {
-	foo := influxdb.NewStats("server")
-	foo.Set("a", 100)
-	foo.Set("b", 600)
+	s := influxdb.NewStats("server")
+	s.Set("a", 100)
+	s.Set("b", 600)
 
-	bar := foo.Snapshot()
-	if bar.Name() != "server" || bar.Get("a") != 100 || bar.Get("b") != 600 {
-		t.Fatalf("stats snapshot returned unexpected result: %#v", bar)
+	other := s.Clone()
+	if other.Name() != "server" || other.Get("a") != 100 || other.Get("b") != 600 {
+		t.Fatalf("stats snapshot returned unexpected result: %#v", other)
 	}
 }
 
 func TestStats_String(t *testing.T) {
-	foo := influxdb.NewStats("server")
-	foo.Set("a", 100)
-	foo.Set("b", 600)
+	s := influxdb.NewStats("server")
+	s.Set("a", 100)
+	s.Set("b", 600)
 
-	if exp, got := `{"server":[{"a":100},{"b":600}]}`, foo.String(); exp != got {
+	if exp, got := `{"server":[{"a":100},{"b":600}]}`, s.String(); exp != got {
 		t.Log("exp: ", exp)
 		t.Log("got: ", got)
 		t.Fatalf("failed to get string")
 	}
-}
-
-// TestStats_RaceCheck is meant to be run with race-detection enabled.
-func TestStats_RaceCheck(t *testing.T) {
-	if testing.Short() {
-		t.Skip("skipping long-runnning Stats race checking")
-	}
-
-	foo := influxdb.NewStats("server")
-	foo.Set("a", 100)
-
-	walker := func(s string, i int64) {
-		if i == -1 {
-			return // Will never happen.
-		}
-	}
-
-	go func() {
-		for {
-			foo.Add("a", 1)
-		}
-	}()
-
-	go func() {
-		for {
-			foo.Walk(walker)
-		}
-	}()
-
-	time.Sleep(2 * time.Second)
-	t.Log("TestStats_RaceCheck completed")
 }
